[sandbox][api] Use new String::Write* APIs
In https://crrev.com/c/5975682 we introduced new version of the
String::Write* APIs. In this CL, all users of the old API in V8 are
converted to the new API.

Bug: 373485796, 376071292
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5979449
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97172}

+++ include/v8-primitive.h
@@ -624,7 +624,7 @@ class V8_EXPORT String : public Name {
 
    private:
     char* str_;
-    int length_;
+    size_t length_;
   };
 
   /**
@@ -654,7 +654,7 @@ class V8_EXPORT String : public Name {
 
    private:
     uint16_t* str_;
-    int length_;
+    uint32_t length_;
   };
 
   /**
